Refactor interface_http.Client.on_relation_changed
This change attempts to correctly organize the logic needed by the
interface to get a Service resource from k8s. The logic has been moved
away from the FrameworkAdapter and into a get_service_spec() function in
the k8s adapter module where it makes more sense. There is still much to
be improved on in the interface_http module though. Specifically on how
the classes are organized so that they are less coupled. However, I will
save that work for a later time.