Support relation to MySQL or MariaDB charm and persistent SQLite store
This change allows a charm user to relate Grafana to a MySQL or MariaDB
charm that provides a 'mysql' interface and a data structure with fields
as used by interface_mysql.MySQLServerDetails. This is the first step in
making Grafana truly HA. Further tests need to be added to ensure that
multiple Grafana pods connect succesfully to the same database cluster.
Failover tests are not necessary since all Grafana pods will be active
and automatically fronted by a k8s Service object which appears to be
the default behavior of Juju CAAS in v2.7.x

This change also adds a minor addition to metadata.yaml such that a
persistent volume is mounted on the default path with the SQLite
database would be created. That way, when a single pod Grafana that is
not connected to an external database is restarted, its data will be
retained when it comes back.